fix(resources): iframe scaling failed on mobile screens with mobile UA
The bug is due to two things:
1. Wix delivers a mobile-optimized website that caps width to 320px when they receive a mobile UA. Our code assumes width to be 1440px across all screen sizes/UA, which fails. The fix is to auto-detect UA client-side and set the IFRAME_DEFAULTS to 320 or 1440 for mobile/desktop respectively.

2. `calculateResponsiveScale` caps the return value via `Math.min(containerWidth / sourceWidth, 1)`. So when sourceWidth is set to 320, and the code passes user screen width to containerWidth, it still does not work since mobile screens are usually 450px => 450/320 > 1 -> 1 (capped). The fix is to remove the Math.min().

These are the main fixes of the code. This commit also incorporates some usual cleanup of the code, e.g., inlining functions/code that are only called once (e.g., calculateResponsiveScale, createScaleTransformStyle). Also, it removes the unnecessary <div> that wraps the <iframe>, and instead applies styles directly on the iframe.