fix(resources): apply CSS transform trickery only to Wixsite
This commit should resolve most of the questions of why we need so much code for a single iframe scenario, as we typically assume all websites adhere to responsive designs. To be fair, WordPress link we received does, but just not for the wixsite link. Unfortunately, we kind of entered a local maximum and decided to fix everything ourselves (a wiser way would be asking the wixsite owner for the account and optimizing on that side).

Therefore, we disable the tricks on the WordPress website, as it is responsive by design.

Left a TODO to explain this big pile of code/logic just for this very edge-case scenario.